requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
gunicorn==20.0.4
appdirs==1.4.4
asgiref==3.3.4
autopep8==1.5.7
boto3==1.17.73
botocore==1.20.73
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
Django==3.2
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
idna==2.10
jmespath==0.10.0
Pillow==8.2.0
# psycopg2==2.8.6 
psycopg2-binary~=2.8.6
pycodestyle==2.7.0
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
virtualenv==20.4.6
virtualenvwrapper-win==1.2.6

